+++ packages/client-ledger/src/sdk/transactions.ts
@@ -13,6 +13,7 @@ import {
 import { HTTPClient } from "../lib/http.js";
 import * as schemas$ from "../lib/schemas.js";
 import { ClientSDK, RequestOptions } from "../lib/sdks.js";
+import { extractSecurity } from "../lib/security.js";
 import * as errors from "../models/errors/index.js";
 import * as operations from "../models/operations/index.js";
 import { createPageIterator, PageIterator, Paginator } from "../types/operations.js";
@@ -83,11 +84,7 @@ export class Transactions extends ClientSDK {
             Accept: "application/json",
         });
 
-        const security$ =
-            typeof this.options$.security === "function"
-                ? await this.options$.security()
-                : this.options$.security;
-
+        const security$ = await extractSecurity(this.options$.security);
         const context = {
             operationID: "Transactions_count",
             oAuth2Scopes: ["ledger:read"],
@@ -175,11 +172,7 @@ export class Transactions extends ClientSDK {
             Accept: "application/json",
         });
 
-        const security$ =
-            typeof this.options$.security === "function"
-                ? await this.options$.security()
-                : this.options$.security;
-
+        const security$ = await extractSecurity(this.options$.security);
         const context = {
             operationID: "Transactions_list",
             oAuth2Scopes: ["ledger:read"],
@@ -298,11 +291,7 @@ export class Transactions extends ClientSDK {
             }),
         });
 
-        const security$ =
-            typeof this.options$.security === "function"
-                ? await this.options$.security()
-                : this.options$.security;
-
+        const security$ = await extractSecurity(this.options$.security);
         const context = {
             operationID: "Transactions_create",
             oAuth2Scopes: ["ledger:write"],
@@ -394,11 +383,7 @@ export class Transactions extends ClientSDK {
             Accept: "application/json",
         });
 
-        const security$ =
-            typeof this.options$.security === "function"
-                ? await this.options$.security()
-                : this.options$.security;
-
+        const security$ = await extractSecurity(this.options$.security);
         const context = {
             operationID: "Transactions_get",
             oAuth2Scopes: ["ledger:read"],
@@ -486,11 +471,7 @@ export class Transactions extends ClientSDK {
             }),
         });
 
-        const security$ =
-            typeof this.options$.security === "function"
-                ? await this.options$.security()
-                : this.options$.security;
-
+        const security$ = await extractSecurity(this.options$.security);
         const context = {
             operationID: "Transactions_addMetadata",
             oAuth2Scopes: ["ledger:write"],
@@ -578,11 +559,7 @@ export class Transactions extends ClientSDK {
             Accept: "application/json",
         });
 
-        const security$ =
-            typeof this.options$.security === "function"
-                ? await this.options$.security()
-                : this.options$.security;
-
+        const security$ = await extractSecurity(this.options$.security);
         const context = {
             operationID: "Transactions_deleteMetadata",
             oAuth2Scopes: ["ledger:write"],
@@ -674,11 +651,7 @@ export class Transactions extends ClientSDK {
             Accept: "application/json",
         });
 
-        const security$ =
-            typeof this.options$.security === "function"
-                ? await this.options$.security()
-                : this.options$.security;
-
+        const security$ = await extractSecurity(this.options$.security);
         const context = {
             operationID: "Transactions_revert",
             oAuth2Scopes: ["ledger:write"],